refactor(config): 디렉토리 위치변경 및 통합
