+++ src/test/java/com/pagoda/matchmeal/controller/FoodControllerTest.java
@@ -1,7 +1,7 @@
 package com.pagoda.matchmeal.controller;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.pagoda.matchmeal.config.jwt.JwtTokenProvider;
+import com.pagoda.matchmeal.common.config.jwt.JwtTokenProvider;
 import com.pagoda.matchmeal.model.dto.UserDto;
 import com.pagoda.matchmeal.model.dto.request.FoodRequestDto;
 import com.pagoda.matchmeal.service.FoodService;
